control, unlink device

+++ SmartHome/app/device/api.py
@@ -8,6 +8,9 @@ from app.device.schemas import AddDeviceSchema, EditDeviceSchema, DeviceSchema
 from app.device.CRUD import add_device, edit_device, delete_device, get_device, get_all_device
 from app.exceptions.exceptions_user import AccessRightsErrorException
 from app.device.options.options import get_option, OptionsDevice
+from app.device.set_value import set_value
+from app.device.schemas import ConsctionStatusForm
+from app.device.polling import device_polling_edit
 
 
 router = APIRouter(
@@ -22,10 +25,10 @@ logger = logging.getLogger(__name__)
 # async def add_device_url(data:AddDeviceSchema, auth_data: TokenData = Depends(token_dep)):
 
 @router.post("")
-async def add_device_url(data:AddDeviceSchema):
+async def add_device_url(data:AddDeviceSchema, auth_data: TokenData = Depends(token_dep)):
 	try:
-		# if auth_data.user_level != UserLevel.ADMIN:
-		# 	raise AccessRightsErrorException()
+		if auth_data.user_level != UserLevel.ADMIN:
+			raise AccessRightsErrorException()
 		await add_device(data)
 		return "ok"
 	except Exception as e:
@@ -35,10 +38,10 @@ async def add_device_url(data:AddDeviceSchema):
 		return JSONResponse(status_code=400, content=str(e))
 	
 @router.put("/{system_name}")
-async def edit_dev(system_name:str, data:EditDeviceSchema):
+async def edit_dev(system_name:str, data:EditDeviceSchema, auth_data: TokenData = Depends(token_dep)):
 	try:
-		# if auth_data.user_level != UserLevel.ADMIN:
-		# 	raise AccessRightsErrorException()
+		if auth_data.user_level != UserLevel.ADMIN:
+			raise AccessRightsErrorException()
 		await edit_device(system_name, data)
 		return "ok"
 	except Exception as e:
@@ -46,10 +49,10 @@ async def edit_dev(system_name:str, data:EditDeviceSchema):
 		return JSONResponse(status_code=400, content=str(e))
 	
 @router.delete("/{system_name}")
-async def delete_dev(system_name:str):
+async def delete_dev(system_name:str, auth_data: TokenData = Depends(token_dep)):
 	try:
-		# if auth_data.user_level != UserLevel.ADMIN:
-		# 	raise AccessRightsErrorException()
+		if auth_data.user_level != UserLevel.ADMIN:
+			raise AccessRightsErrorException()
 		await delete_device(system_name)
 		return "ok"
 	except Exception as e:
@@ -57,10 +60,10 @@ async def delete_dev(system_name:str):
 		return JSONResponse(status_code=400, content=str(e))
 	
 @router.get("/options", response_model=List[OptionsDevice])
-async def get_options_dev():
+async def get_options_dev(auth_data: TokenData = Depends(token_dep)):
 	try:
-		# if auth_data.user_level != UserLevel.ADMIN:
-		# 	raise AccessRightsErrorException()
+		if auth_data.user_level != UserLevel.ADMIN:
+			raise AccessRightsErrorException()
 		data = get_option()
 		print(data)
 		return data
@@ -69,10 +72,10 @@ async def get_options_dev():
 		return JSONResponse(status_code=400, content=str(e))
 	
 @router.get("/{system_name}", response_model=DeviceSchema)
-async def get_dev(system_name:str):
+async def get_dev(system_name:str, auth_data: TokenData = Depends(token_dep)):
 	try:
-		# if auth_data.user_level == UserLevel.ADMIN:
-		# 	raise AccessRightsErrorException()
+		if auth_data.user_level == UserLevel.NONE:
+			raise AccessRightsErrorException()
 		data = get_device(system_name)
 		return data
 	except Exception as e:
@@ -80,10 +83,10 @@ async def get_dev(system_name:str):
 		return JSONResponse(status_code=400, content=str(e))
 	
 @router.get("", response_model=List[DeviceSchema])
-async def get_all_dev():
+async def get_all_dev(auth_data: TokenData = Depends(token_dep)):
 	try:
-		# if auth_data.user_level == UserLevel.ADMIN:
-		# 	raise AccessRightsErrorException()
+		if auth_data.user_level == UserLevel.NONE:
+			raise AccessRightsErrorException()
 		data = await get_all_device()
 		return data
 	except Exception as e:
@@ -91,3 +94,24 @@ async def get_all_dev():
 		return JSONResponse(status_code=400, content=str(e))
 	
 
+@router.get("/{system_name}/value/{field_name}/set/{value}")
+async def set_device_state(system_name, field_name, value, auth_data: TokenData = Depends(token_dep)):
+	try:
+		if auth_data.user_level == UserLevel.NONE:
+			raise AccessRightsErrorException()
+		await set_value(system_name, field_name, value)
+	except Exception as e:
+		logger.warning(str(e))
+		return JSONResponse(status_code=400, content=str(e))
+
+	
+@router.post("/{system_name}/polling")
+async def set_connection_status(system_name, data:ConsctionStatusForm, auth_data: dict = Depends(token_dep)):
+	try:
+		if auth_data.user_level != UserLevel.ADMIN:
+			raise AccessRightsErrorException()
+		await device_polling_edit(system_name, data.status)
+		return "ok"
+	except Exception as e:
+		logger.warning(str(e))
+		return JSONResponse(status_code=400, content=str(e))
